+++ src/test/java/org/jsapar/Text2StringConverterTest.java
@@ -1,6 +1,6 @@
 package org.jsapar;
 
-import org.jsapar.parse.text.TextParseConfig;
+import org.jsapar.text.TextParseConfig;
 import org.jsapar.schema.CsvSchema;
 import org.jsapar.schema.CsvSchemaCell;
 import org.jsapar.schema.CsvSchemaLine;