+++ src/test/java/org/jsapar/text/EnumFormatTest.java
@@ -1,5 +1,6 @@
 package org.jsapar.text;
 
+import org.jsapar.utils.text.EnumFormat;
 import org.junit.Test;
 
 import java.text.ParseException;